started cleaning training pipeline
